Add ./autogen.sh and 'make dist' target for Jenkins
The normal way to build mirrorbits is to check it out to
$GOPATH/src/github.com/etix/mirrorbits, use govendor to fetch its
dependencies into vendor/, then build it. We would prefer to include all
the dependencies in the source tarball sent to OBS, so we'll run
govendor in Jenkins where the "orig" tarball + Debian source package is
made.

There's a fly hidden in the first step of that ointment: govendor *must*
be run from $GOPATH/src/$PACKAGE, but we don't have any control over
where Jenkins checks the tree out. To work around this, add a little
wrapper script that creates a temporary directory to use as $GOPATH and
symlinks the source directory to the appropriate path within it, and use
it where needed in the Makefile. (While go build does not strictly need
to be run from that path, since mirrorbits refers to parts of its own
code by URL, there does need to be a copy of mirrorbits in the right
place, so in practice we need the same trick there.)

autogen.sh is needed to make Jenkins believe this is a project it can
build with ./autogen.sh && make && make dist; this makes it a suitable
place to enable the USE_TEMP_GOPATH flag. Makefile.inc is included in
the source tarball, so the OBS build will inherit this.

We also add a DISABLE_GOVENDOR flag that will be used on OBS to skip
govendor entirely.

I promise this was easier than packaging Mirrorbits' dependencies
"properly"!

https://phabricator.endlessm.com/T23123