with-temp-gopath.sh: register cleanup after defining $GOPATH
$GOPATH may point to a real, valuable directory on a development system.
It would be sad to wipe out the developer's real $GOPATH if the script
exits between the trap and the mktemp. (Is this actually possible,
given there are no intervening lines? I don't know.)

If we register the trap later, cleanup can be simpler, at the cost of
leaking an empty temporary directory if the script receives SIGINT after
executing mktemp and assigning to GOPATH but before evaluating trap. Oh
well.