.service: send SIGQUIT to stop
The upstream version of this file uses /bin/kill -QUIT $MAINPID but that
caused spurious errors if the service didn't come up properly in the
first place. I misread it as kill -TERM and removed it. Mirrorbits has
special handling of QUIT that waits 5 seconds for running tasks to
finish, so it's worth doing.