autogen.sh: Initialize GOPATH to /usr/share/gocode
Debian installs go packages in /usr/share/gocode, so that needs to be
included in GOPATH. Do the dirty thing and just set it in the generated
Makefile.inc since there's no other clean way to inject environment
variables into our build process.

When trying to figure out how debian normally handles this, it seems
that the go buildsystem in dh-golang makes a symlink farm in
$builddir/src to /usr/share/gocode/src and sets GOPATH to the build
directory. Maybe something to consider in the future.

https://phabricator.endlessm.com/T25314