Revert "mirrorlist: put at least one element in the piechart"
This reverts commit 26a3f963d1eaba11741044e8dc285fca34ff01e6. The
feature to overwrite a variable within a conditional using `=` is new in
Go 1.11[1]. The Go version on bionic is 1.10, so this fails at runtime:

  template: mirrorlist.html:27: unexpected "=" in operand

Since we have enabled mirrors, then I don't think we'll run into the
issue this was fixing anyways.

The revert needs a little adjustment since the label was changed from
the ID attribute to the Name attribute since this commit.

1. https://github.com/golang/go/issues/10608

https://phabricator.endlessm.com/T25314